Rename result modules to snake_case
