+++ src/ffbb_api_client_v2/meilisearch_client.py
@@ -5,7 +5,7 @@ from requests_cache import CachedSession
 from .models.http_requests_helper import catch_result, default_cached_session
 from .models.http_requests_utils import http_post_json
 from .models.multi_search_query import MultiSearchQuery
-from .models.MultiSearchResults import (
+from .models.multi_search_results import (
     MultiSearchResults,
     multi_search_results_from_dict,
 )
