+++ src/ffbb_api_client_v2/meilisearch_client_extension.py
@@ -5,7 +5,7 @@ from requests_cache import CachedSession
 from .meilisearch_client import MeilisearchClient
 from .models.http_requests_helper import default_cached_session
 from .models.multi_search_query import MultiSearchQuery
-from .models.MultiSearchResults import MultiSearchResults
+from .models.multi_search_results import MultiSearchResults
 
 
 class MeilisearchClientExtension(MeilisearchClient):
