+++ src/ffbb_api_client_v2/models/competitions_multi_search_result.py
