+++ src/ffbb_api_client_v2/models/organismes_multi_search_result.py
