+++ src/ffbb_api_client_v2/models/pratiques_multi_search_result.py
