+++ src/ffbb_api_client_v2/models/rencontres_multi_search_result.py
