+++ src/ffbb_api_client_v2/models/salles_multi_search_result.py
