+++ src/ffbb_api_client_v2/models/terrains_multi_search_result.py
