+++ src/ffbb_api_client_v2/models/tournois_multi_search_result.py
